ECOCKPIT-12 irqchip/gic: change affinity settings strategy
Instead of setting the same affinity for all interrupts at boot time,
only set it when interrupts are registered.
This way, a cluster will not mess up the other cluster's settings.

ECOCKPIT-241 do not configure GIC Distributor if already done.
When enabling the group1 irq, the distributor is disabled,
which is causing lost irq to other domain already running.

ECOCKPIT-251 GIC: do not disable SPI IRQs when setting their affinity
disabling the IRQ in gic_set_type() may disturb drivers
that do not 're-enable' it after.

Signed-off-by: Seb Fagard <sebastien.fagard@nxp.com>